+++ Project1-3530/PSLL.h
@@ -230,9 +230,8 @@ namespace cop3530 {
 		//#done
 
 		//checks for size
-		int size() const { return amount; }
+		size_t size() const { return amount; }
 		//#done
-
 		//clears list
 		void clear() {
 			Node * temp = head;
